+++ src/index.ts
@@ -1,2 +1,3 @@
 export * from "./ReactDecorators.js";
 export * from "./ReactRoute.js";
+export * from "./static.js";